Fix sign comparison warnings

+++ src/apps/atlas-benchmark.cc
@@ -222,7 +222,7 @@ private:
   size_t nlev;
   size_t N;
   size_t niter;
-  long exclude;
+  size_t exclude;
   bool output;
   long omp_threads;
   double dz;
@@ -275,7 +275,7 @@ void AtlasBenchmark::run()
     Log::info() << "      |----|----|----|----|----|----|----|----|----|----|\n";
     Log::info() << "      " << std::flush;
   }
-  int tic=0;
+  unsigned int tic=0;
   for( iter=0; iter<niter; ++iter )
   {
     if( progress )
@@ -357,7 +357,7 @@ void AtlasBenchmark::setup()
   double radius = 6371.22e+03; // Earth's radius
   double height = 80.e+03;     // Height of atmosphere
   double deg2rad = M_PI/180.;
-  for( int jnode=0; jnode<nnodes; ++jnode )
+  for(size_t jnode = 0; jnode < nnodes; ++jnode)
   {
     lonlat(jnode,LON) = lonlat(jnode,LON) * deg2rad;
     lonlat(jnode,LAT) = lonlat(jnode,LAT) * deg2rad;
@@ -367,10 +367,10 @@ void AtlasBenchmark::setup()
     double G  = hx*hy;
     V(jnode) *= std::pow(deg2rad,2) * G;
 
-    for( int jlev=0; jlev<nlev; ++jlev )
+    for(size_t jlev = 0; jlev < nlev; ++jlev)
       field(jnode,jlev) = 100.+50.*std::cos(2*y);
   }
-  for( int jedge=0; jedge<nedges; ++jedge )
+  for(size_t jedge = 0; jedge < nedges; ++jedge)
   {
     S(jedge,LON) *= deg2rad;
     S(jedge,LAT) *= deg2rad;
@@ -384,12 +384,12 @@ void AtlasBenchmark::setup()
   node2edge_sign = ArrayView<double,2> ( mesh->function_space("nodes").
                                          create_field<double>("to_edge_sign",node2edge.shape(1)) );
 
-  for( int jnode=0; jnode<nnodes; ++jnode )
+  for(size_t jnode = 0; jnode < nnodes; ++jnode)
   {
-    for( int jedge=0; jedge<node2edge_size(jnode); ++jedge )
+    for(int jedge = 0; jedge < node2edge_size(jnode); ++jedge)
     {
       int iedge = node2edge(jnode,jedge);
-      int ip1 = edge2node(iedge,0);
+      size_t ip1 = edge2node(iedge,0);
       if( jnode == ip1 )
         node2edge_sign(jnode,jedge) = 1.;
       else
@@ -399,7 +399,7 @@ void AtlasBenchmark::setup()
 
   vector<int> tmp(nedges);
   int c(0);
-  for( int jedge=0; jedge<nedges; ++jedge )
+  for(size_t jedge = 0; jedge < nedges; ++jedge)
   {
     if( edge_is_pole(jedge) )
       tmp[c++] = jedge;
@@ -410,7 +410,7 @@ void AtlasBenchmark::setup()
 
   ArrayView<int,1> flags( mesh->function_space("nodes").field("flags") );
   is_ghost.reserve(nnodes);
-  for( int jnode=0; jnode<nnodes; ++jnode )
+  for(size_t jnode = 0; jnode < nnodes; ++jnode)
   {
     is_ghost.push_back( Topology::check(flags(jnode),Topology::GHOST) );
   }
@@ -441,12 +441,12 @@ void AtlasBenchmark::iteration()
 #ifdef ATLAS_HAVE_OMP
   #pragma omp parallel for
 #endif
-  for( int jedge=0; jedge<nedges; ++jedge )
+  for(size_t jedge = 0; jedge < nedges; ++jedge)
   {
     int ip1 = edge2node(jedge,0);
     int ip2 = edge2node(jedge,1);
 
-    for( int jlev=0; jlev<nlev; ++jlev )
+    for(size_t jlev = 0; jlev < nlev; ++jlev)
     {
       double avg = ( field(ip1,jlev) + field(ip2,jlev) ) * 0.5;
       avgS(jedge,jlev,LON) = S(jedge,LON)*avg;
@@ -457,9 +457,9 @@ void AtlasBenchmark::iteration()
 #ifdef ATLAS_HAVE_OMP
   #pragma omp parallel for
 #endif
-  for( int jnode=0; jnode<nnodes; ++jnode )
+  for(size_t jnode = 0; jnode < nnodes; ++jnode)
   {
-    for( int jlev=0; jlev<nlev; ++jlev )
+    for(size_t jlev = 0; jlev < nlev; ++jlev )
     {
       grad(jnode,jlev,LON) = 0.;
       grad(jnode,jlev,LAT) = 0.;
@@ -468,13 +468,13 @@ void AtlasBenchmark::iteration()
     {
       int iedge = node2edge(jnode,jedge);
       double add = node2edge_sign(jnode,jedge);
-      for( int jlev=0; jlev<nlev; ++jlev )
+      for(size_t jlev = 0; jlev < nlev; ++jlev)
       {
         grad(jnode,jlev,LON) += add*avgS(iedge,jlev,LON);
         grad(jnode,jlev,LAT) += add*avgS(iedge,jlev,LAT);
       }
     }
-    for( int jlev=0; jlev<nlev; ++jlev )
+    for(size_t jlev = 0; jlev < nlev; ++jlev)
     {
       grad(jnode,jlev,LON) /= V(jnode);
       grad(jnode,jlev,LAT) /= V(jnode);
@@ -482,12 +482,12 @@ void AtlasBenchmark::iteration()
   }
   // special treatment for the north & south pole cell faces
   // Sx == 0 at pole, and Sy has same sign at both sides of pole
-  for( int jedge=0; jedge<pole_edges.size();++jedge )
+  for(size_t jedge = 0; jedge < pole_edges.size(); ++jedge)
   {
     int iedge = pole_edges[jedge];
     int ip2 = edge2node(iedge,1);
     // correct for wrong Y-derivatives in previous loop
-    for( int jlev=0; jlev<nlev; ++jlev )
+    for(size_t jlev = 0; jlev < nlev; ++jlev)
       grad(ip2,jlev,LAT) += 2.*avgS(iedge,jlev,LAT)/V(ip2);
   }
 
@@ -497,11 +497,11 @@ void AtlasBenchmark::iteration()
 #ifdef ATLAS_HAVE_OMP
   #pragma omp parallel for
 #endif
-  for( int jnode=0; jnode<nnodes; ++jnode )
+  for(size_t jnode = 0; jnode < nnodes; ++jnode)
   {
     if( nlev > 2 )
     {
-      for( int jlev=1; jlev<nlev-1; ++jlev )
+      for(size_t jlev = 1; jlev < nlev - 1; ++jlev)
       {
         grad(jnode,jlev,ZZ)   = (field(jnode,jlev+1)     - field(jnode,jlev-1))*dzi_2;
       }
@@ -555,11 +555,11 @@ double AtlasBenchmark::result()
   double maxval = numeric_limits<double>::min();
   double minval = numeric_limits<double>::max();;
   double norm = 0.;
-  for( int jnode=0; jnode<nnodes; ++jnode )
+  for(size_t jnode = 0; jnode < nnodes; ++jnode)
   {
     if( !is_ghost[jnode] )
     {
-      for( int jlev=0; jlev<nlev; ++jlev )
+      for(size_t jlev = 0; jlev < nlev; ++jlev)
       {
         maxval = max(maxval,grad(jnode,jlev,LON));
         maxval = max(maxval,grad(jnode,jlev,LAT));
